requirements.txt
dnspython
kurigram
motor
pymongo
TgCrypto
python-dotenv
flask
dnspython
flask
kurigram
motor
pymongo
python-dotenv
TgCrypto
aiofiles
aiohttp
tgcrypto
uvloop
aiofiles
aiohttp
dnspython
flask
kurigram
motor
pymongo
python-dotenv
TgCrypto
aiofiles
aiohttp
dnspython
flask
kurigram
motor
pymongo
python-dotenv
TgCrypto
uvloop
aiofiles
aiohttp
dnspython
flask
kurigram
motor
pymongo
python-dotenv
TgCrypto
uvloop
tgcrypto
uvloop